% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_devices.R
\name{bc_devices}
\alias{bc_devices}
\title{Get Devices Available for \code{tinygrad}/\code{brainchop}}
\usage{
bc_devices()
}
\value{
A \code{data.frame} similar to \code{tinygrad.device}
}
\description{
Get Devices Available for \code{tinygrad}/\code{brainchop}
}
\examples{
bc_devices()
}
