% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{bc_list_models}
\alias{bc_list_models}
\title{List Brainchop Models}
\usage{
bc_list_models()
}
\value{
A list of models and a printout
}
\description{
List Brainchop Models
}
\examples{
bc_list_models()
}
