% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainchop.R
\name{mindgrab}
\alias{mindgrab}
\alias{brainchop}
\title{Run Brain Chop/Mind Grab}
\usage{
mindgrab(
  input,
  model = "mindgrab",
  ct = FALSE,
  comply = FALSE,
  crop = FALSE,
  crop_percent = 2L,
  border = 0L,
  inverse_conform = FALSE,
  export_classes = FALSE,
  device = c("CPU", "NV", "AMD", "QCOM", "METAL", "CUDA", "GPU", "LLVM", "WEBGPU")
)

brainchop(
  input,
  model = "mindgrab",
  ct = FALSE,
  comply = FALSE,
  crop = FALSE,
  crop_percent = 2L,
  border = 0L,
  inverse_conform = FALSE,
  export_classes = FALSE,
  device = c("CPU", "NV", "AMD", "QCOM", "METAL", "CUDA", "GPU", "LLVM", "WEBGPU")
)
}
\arguments{
\item{input}{input image or filename}

\item{model}{Model name to use, see \code{\link[=bc_list_models]{bc_list_models()}}, current options are
"mindgrab", "tissue_fast", "subcortical", "DKatlas", "multiaxial".}

\item{ct}{Is the image a CT scan? (defaults to FALSE)}

\item{comply}{Insert compliance arguments to \code{niimath} before '-conform'}

\item{crop}{Crop the input for faster execution. May reduce accuracy.(defaults to percentile 2 cutoff)}

\item{crop_percent}{Crop percent cutoff for the input image, defaults to 2L.}

\item{border}{Mask border threshold in mm. Default is 0. Makes a difference
only if the model is \code{mindgrab}}

\item{inverse_conform}{Perform inverse conformation into original image space}

\item{export_classes}{Export class probability maps}

\item{device}{device for running the model, defaults to "CPU". If
\code{NULL}, it will use the default device set in the environment.}
}
\value{
A list of the input and output images and files created,
including the mask.
}
\description{
Run Brain Chop/Mind Grab
}
